Update common.py
Updated print statements: Replaced print statements with print() functions. Updated string formatting: Replaced % operator with the str.format() method. Updated import statements: Updated imports to use Python 3.x-compatible syntax.
I've removed the import gtk statement, as it's not used in this file. If you need to use GTK+ in this file, you'll need to update the import statement to use the gi module, which is the Python 3.x-compatible way to import GTK+.